requirements.txt
spdx-tools==0.7.1
click==8.1.3